fix: add docs and clean code
