+++ packages/blocks/src/root-block/widgets/format-bar/config.ts
@@ -1,6 +1,5 @@
 import type { Chain, InitCommandCtx } from '@blocksuite/block-std';
 import { assertExists } from '@blocksuite/global/utils';
-import type { BlockModel, Doc } from '@blocksuite/store';
 import { html, type TemplateResult } from 'lit';
 
 import { toast } from '../../../_common/components/index.js';
@@ -28,7 +27,7 @@ import {
   TextIcon,
   UnderlineIcon,
 } from '../../../_common/icons/index.js';
-import { matchFlavours } from '../../../_common/utils/index.js';
+import { createLinkedDocFromSelectedBlocks } from '../../../_common/utils/render-linked-doc.js';
 import type { AffineFormatBarWidget } from './format-bar.js';
 
 export type DividerConfigItem = {
@@ -179,84 +178,16 @@ export function toolbarDefaultConfig(toolbar: AffineFormatBarWidget) {
           .run();
         const { selectedModels } = ctx;
         assertExists(selectedModels);
+        if (!selectedModels.length) return;
 
         const host = formatBar.host;
         host.selection.clear();
 
         const doc = host.doc;
-        const linkedDoc = doc.collection.createDoc({});
-        linkedDoc.load(() => {
-          const rootId = linkedDoc.addBlock('affine:page', {
-            title: new doc.Text(''),
-          });
-          linkedDoc.addBlock('affine:surface', {}, rootId);
-          const noteId = linkedDoc.addBlock('affine:note', {}, rootId);
-
-          const firstBlock = selectedModels[0];
-          assertExists(firstBlock);
-
-          doc.addSiblingBlocks(
-            firstBlock,
-            [
-              {
-                flavour: 'affine:embed-linked-doc',
-                pageId: linkedDoc.id,
-              },
-            ],
-            'before'
-          );
-
-          if (
-            matchFlavours(firstBlock, ['affine:paragraph']) &&
-            firstBlock.type.match(/^h[1-6]$/)
-          ) {
-            const title = firstBlock.text.toString();
-            linkedDoc.collection.setDocMeta(linkedDoc.id, {
-              title,
-            });
-
-            const linkedDocRootModel = linkedDoc.getBlockById(rootId);
-            assertExists(linkedDocRootModel);
-            linkedDoc.updateBlock(linkedDocRootModel, {
-              title: new doc.Text(title),
-            });
-
-            doc.deleteBlock(firstBlock);
-            selectedModels.shift();
-          }
-
-          const addBlockRecursively = (
-            linkedDoc: Doc,
-            model: BlockModel,
-            parentId: string
-          ) => {
-            // Add current block to linked doc
-            const keys = model.keys as (keyof typeof model)[];
-            const values = keys.map(key => model[key]);
-            const blockProps = Object.fromEntries(
-              keys.map((key, i) => [key, values[i]])
-            );
-            const newModelId = linkedDoc.addBlock(
-              model.flavour as never,
-              blockProps,
-              parentId
-            );
-            // Add children to linked doc, parent is the new model
-            const children = model.children;
-            if (children.length > 0) {
-              children.forEach(child => {
-                addBlockRecursively(linkedDoc, child, newModelId);
-              });
-            }
-            // Delete current block from original doc
-            doc.deleteBlock(model);
-          };
-          // Add selected blocks to linked doc recursively
-          selectedModels.forEach(model => {
-            addBlockRecursively(linkedDoc, model, noteId);
-          });
-        });
-
+        const linkedDoc = createLinkedDocFromSelectedBlocks(
+          doc,
+          selectedModels
+        );
         const linkedDocService = host.spec.getService(
           'affine:embed-linked-doc'
         );
